Fix checker from crashing when package has multiple licenses
Packages like json-schema@0.2.3 can return an array of licenses from davglass/license-checker - in this case [ 'AFLv2.1', 'BSD' ]

This causes license-checker-ci to crash as follows: ERROR: Uncaught exception. Details: {}

A better fix would be to check all licenses in these arrays, but I'm happy for now just to have it fail the check and need to add it to manualOverrides. Unless you want to make it smarter.